+++ src/test/java/study/step_1/SetTest.java
@@ -1,4 +1,4 @@
-package study;
+package study.step_1;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
