+++ src/test/java/study/step_1/StringTest.java
@@ -1,4 +1,4 @@
-package study;
+package study.step_1;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
