+++ src/test/java/study/step_2/CalculatorTest.java
@@ -1,4 +1,4 @@
-package study;
+package study.step_2;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
