+++ src/test/java/study/step_2/Operator.java
@@ -1,4 +1,4 @@
-package study;
+package study.step_2;
 
 import java.util.function.BiFunction;
 import java.util.stream.Stream;
